server/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.12.3
python-dotenv==1.0.1
httpx==0.27.2
typing-extensions==4.14.1
convex==0.7.0
google-generativeai==0.8.3
anthropic==0.40.0
openai==1.58.1